pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "python-disposable"
dynamic = ["version"]
description = "Check if an email address or domain belongs to a disposable/temporary email service. Bundles 72k+ domains from disposable/disposable-email-domains."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["email", "disposable", "validation", "spam"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
test = ["pytest>=7"]
osint = [
    "dnspython>=2.4",
]

[project.urls]
"Homepage" = "https://github.com/pescheckit/python-disposable"
"Source" = "https://github.com/pescheckit/python-disposable"
"Issues" = "https://github.com/pescheckit/python-disposable/issues"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "disposable_email/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["disposable_email"]
# Ship code + the small text data files; exclude the OSINT working artifacts
# (multi-MB SQLite, intermediate candidates) — those stay in the repo for
# transparency but bloat the wheel for end users who only need the lookup API.
exclude = [
    "disposable_email/data/resolution.sqlite",
    "disposable_email/data/candidates.txt",
]

[tool.hatch.build.targets.sdist]
include = [
    "disposable_email/",
    "scripts/",
    "docs/",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
exclude = [
    "disposable_email/data/resolution.sqlite",
]
